requirements.txt
Pillow
InstagramApi
schedule
